Reorganize the TSO test and reduce the integration test time cost
